+++ packages/atomic/cypress/integration/standalone-search-box.cypress.ts
@@ -1,11 +1,11 @@
 import {TestFixture} from '../fixtures/test-fixture';
-import {addSearchBox} from './search-box-actions';
+import {addSearchBox} from './search-box/search-box-actions';
+import {SearchBoxSelectors} from './search-box/search-box-selectors';
 import {
   assertHasText,
-  assertLogSearchFromLink,
   assertLogOmniboxFromLink,
-} from './search-box-assertions';
-import {SearchBoxSelectors} from './search-box-selectors';
+  assertLogSearchFromLink,
+} from './search-box/search-box-assertions';
 
 describe('Standalone Search Box Test Suites', () => {
   function setupStandaloneSearchBox(url = '/test.html') {