+++ packages/atomic/src/components.d.ts
@@ -9,6 +9,7 @@ import { CategoryFacetSortCriterion, DateFilter, DateFilterState, FacetSortCrite
 import { Bindings } from "./utils/initialization-utils";
 import { NumberInputType } from "./components/facets/facet-number-input/number-input-type";
 import { ResultDisplayDensity, ResultDisplayImageSize, ResultDisplayLayout } from "./components/atomic-result/atomic-result-display-options";
+import { ResultRenderingFunction } from "./components/result-lists/result-list-common";
 import { Section } from "./components/atomic-layout-section/sections";
 import { ObservableMap } from "@stencil/store";
 import { AtomicStore } from "./utils/store";
@@ -248,8 +249,10 @@ export namespace Components {
         "density": ResultDisplayDensity;
         /**
           * A list of non-default fields to include in the query results, separated by commas.
+          * @deprecated add it to atomic-search-interface instead
          */
         "fieldsToInclude": string;
+        "focusOnNextNewResult": () => Promise<void>;
         /**
           * The expected size of the image displayed in the results.
          */
@@ -263,7 +266,7 @@ export namespace Components {
           * Sets a rendering function to bypass the standard HTML template mechanism for rendering results. You can use this function while working with web frameworks that don't use plain HTML syntax, e.g., React, Angular or Vue.  Do not use this method if you integrate Atomic in a plain HTML deployment.
           * @param render
          */
-        "setRenderFunction": (render: (result: FoldedResult) => HTMLElement) => Promise<void>;
+        "setRenderFunction": (render: ResultRenderingFunction) => Promise<void>;
     }
     interface AtomicFormatCurrency {
         /**
@@ -567,7 +570,7 @@ export namespace Components {
         /**
           * The result content to display.
          */
-        "content": ParentNode;
+        "content"?: ParentNode;
         /**
           * How large or small results should be.
          */
@@ -589,6 +592,10 @@ export namespace Components {
          */
         "imageSize"?: ResultDisplayImageSize;
         "loadingFlag"?: string;
+        /**
+          * Internal function used by atomic-result-list in advanced setup, that allows to bypass the standard HTML template system. Particularly useful for Atomic React
+         */
+        "renderingFunction"?: ResultRenderingFunction;
         /**
           * The result item.
          */
@@ -690,8 +697,10 @@ export namespace Components {
         "display": ResultDisplayLayout;
         /**
           * A list of non-default fields to include in the query results, separated by commas.
+          * @deprecated add it to atomic-search-interface instead
          */
         "fieldsToInclude": string;
+        "focusOnNextNewResult": () => Promise<void>;
         /**
           * @deprecated use `imageSize` instead.
          */
@@ -704,7 +713,7 @@ export namespace Components {
           * Sets a rendering function to bypass the standard HTML template mechanism for rendering results. You can use this function while working with web frameworks that don't use plain HTML syntax, e.g., React, Angular or Vue.  Do not use this method if you integrate Atomic in a plain HTML deployment.
           * @param render
          */
-        "setRenderFunction": (render: (result: Result) => HTMLElement) => Promise<void>;
+        "setRenderFunction": (render: ResultRenderingFunction) => Promise<void>;
     }
     interface AtomicResultMultiValueText {
         /**
@@ -712,7 +721,7 @@ export namespace Components {
          */
         "delimiter": string | null;
         /**
-          * The field that the component should use. The component will try to find this field in the `Result.raw` object unless it finds it in the `Result` object first. Make sure this field is present in the `fieldsToInclude` property of the `atomic-result-list` component.
+          * The field that the component should use. The component will try to find this field in the `Result.raw` object unless it finds it in the `Result` object first. Make sure this field is present in the `fieldsToInclude` property of the `atomic-search-interface` component.
          */
         "field": string;
         /**
@@ -722,7 +731,7 @@ export namespace Components {
     }
     interface AtomicResultNumber {
         /**
-          * The field that the component should use. The component will try to find this field in the `Result.raw` object unless it finds it in the `Result` object first. Make sure this field is present in the `fieldsToInclude` property of the `atomic-result-list` component.
+          * The field that the component should use. The component will try to find this field in the `Result.raw` object unless it finds it in the `Result` object first. Make sure this field is present in the `fieldsToInclude` property of the `atomic-search-interface` component.
          */
         "field": string;
     }
@@ -844,6 +853,15 @@ export namespace Components {
           * The expected size of the image displayed in the results.
          */
         "imageSize": ResultDisplayImageSize;
+        /**
+          * The maximum number of results to show.
+         */
+        "maxResultsPerQuery": number;
+        /**
+          * Sets a rendering function to bypass the standard HTML template mechanism for rendering results. You can use this function while working with web frameworks that don't use plain HTML syntax, e.g., React, Angular or Vue.  Do not use this method if you integrate Atomic in a plain HTML deployment.
+          * @param render
+         */
+        "setRenderFunction": (render: ResultRenderingFunction) => Promise<void>;
     }
     interface AtomicSearchBoxQuerySuggestions {
         /**
@@ -886,6 +904,10 @@ export namespace Components {
           * Executes the first search and logs the interface load event to analytics, after initializing connection to the headless search engine.
          */
         "executeFirstSearch": () => Promise<void>;
+        /**
+          * A list of non-default fields to include in the query results, separated by commas.
+         */
+        "fieldsToInclude": string;
         /**
           * The search interface i18next instance.
          */
@@ -965,7 +987,7 @@ export namespace Components {
         /**
           * The non-localized label for the facet. Used in the `atomic-breadbox` component through the bindings store.
          */
-        "label": string;
+        "label"?: string;
         /**
           * The number of values to request for this facet. Also determines the number of additional values to request each time more values are shown.
          */
@@ -1935,6 +1957,7 @@ declare namespace LocalJSX {
         "density"?: ResultDisplayDensity;
         /**
           * A list of non-default fields to include in the query results, separated by commas.
+          * @deprecated add it to atomic-search-interface instead
          */
         "fieldsToInclude"?: string;
         /**
@@ -2251,7 +2274,7 @@ declare namespace LocalJSX {
         /**
           * The result content to display.
          */
-        "content": ParentNode;
+        "content"?: ParentNode;
         /**
           * How large or small results should be.
          */
@@ -2273,6 +2296,10 @@ declare namespace LocalJSX {
          */
         "imageSize"?: ResultDisplayImageSize;
         "loadingFlag"?: string;
+        /**
+          * Internal function used by atomic-result-list in advanced setup, that allows to bypass the standard HTML template system. Particularly useful for Atomic React
+         */
+        "renderingFunction"?: ResultRenderingFunction;
         /**
           * The result item.
          */
@@ -2370,6 +2397,7 @@ declare namespace LocalJSX {
         "display"?: ResultDisplayLayout;
         /**
           * A list of non-default fields to include in the query results, separated by commas.
+          * @deprecated add it to atomic-search-interface instead
          */
         "fieldsToInclude"?: string;
         /**
@@ -2387,7 +2415,7 @@ declare namespace LocalJSX {
          */
         "delimiter"?: string | null;
         /**
-          * The field that the component should use. The component will try to find this field in the `Result.raw` object unless it finds it in the `Result` object first. Make sure this field is present in the `fieldsToInclude` property of the `atomic-result-list` component.
+          * The field that the component should use. The component will try to find this field in the `Result.raw` object unless it finds it in the `Result` object first. Make sure this field is present in the `fieldsToInclude` property of the `atomic-search-interface` component.
          */
         "field": string;
         /**
@@ -2397,7 +2425,7 @@ declare namespace LocalJSX {
     }
     interface AtomicResultNumber {
         /**
-          * The field that the component should use. The component will try to find this field in the `Result.raw` object unless it finds it in the `Result` object first. Make sure this field is present in the `fieldsToInclude` property of the `atomic-result-list` component.
+          * The field that the component should use. The component will try to find this field in the `Result.raw` object unless it finds it in the `Result` object first. Make sure this field is present in the `fieldsToInclude` property of the `atomic-search-interface` component.
          */
         "field": string;
     }
@@ -2515,6 +2543,10 @@ declare namespace LocalJSX {
           * The expected size of the image displayed in the results.
          */
         "imageSize"?: ResultDisplayImageSize;
+        /**
+          * The maximum number of results to show.
+         */
+        "maxResultsPerQuery"?: number;
     }
     interface AtomicSearchBoxQuerySuggestions {
         /**
@@ -2553,6 +2585,10 @@ declare namespace LocalJSX {
           * The search interface headless engine.
          */
         "engine"?: SearchEngine;
+        /**
+          * A list of non-default fields to include in the query results, separated by commas.
+         */
+        "fieldsToInclude"?: string;
         /**
           * The search interface i18next instance.
          */