+++ packages/atomic/src/components/result-lists/grid-display-results.tsx
@@ -6,13 +6,20 @@ import {ResultsProps} from './result-list-common';
 export const GridDisplayResults: FunctionalComponent<ResultsProps> = (
   props
 ) => {
-  return props.resultListState.results.map((result) => {
+  return props.resultListState.results.map((result, index) => {
     const interactiveResult = buildInteractiveResult(props.bindings.engine, {
       options: {result: props.resultListCommon.getUnfoldedResult(result)},
     });
 
     return (
-      <div part="result-list-grid-clickable-container outline">
+      <div
+        part="result-list-grid-clickable-container outline"
+        ref={(element) =>
+          element &&
+          props.indexOfResultToFocus === index &&
+          props.newResultRef?.(element)
+        }
+      >
         <LinkWithResultAnalytics
           part="result-list-grid-clickable"
           onSelect={() => interactiveResult.select()}
@@ -21,6 +28,8 @@ export const GridDisplayResults: FunctionalComponent<ResultsProps> = (
           href={props.resultListCommon.getUnfoldedResult(result).clickUri}
           target="_self"
           title={props.resultListCommon.getUnfoldedResult(result).title}
+          tabIndex={-1}
+          ariaHidden={true}
         />
         <atomic-result
           key={props.resultListCommon.getResultId(
@@ -30,11 +39,9 @@ export const GridDisplayResults: FunctionalComponent<ResultsProps> = (
           result={result}
           engine={props.bindings.engine}
           store={props.bindings.store}
-          display={props.display}
-          density={props.density}
-          imageSize={props.imageSize}
           content={props.getContentOfResultTemplate(result)}
           loadingFlag={props.resultListCommon.loadingFlag}
+          {...props}
         ></atomic-result>
       </div>
     );