+++ packages/atomic/src/global/global.ts
@@ -1,0 +1,7 @@
+import {setCoveoGlobal} from './environment';
+import {loadFocusVisiblePolyfill} from './focus-visible';
+
+export function loadGlobalScripts() {
+  setCoveoGlobal();
+  loadFocusVisiblePolyfill();
+}