+++ packages/atomic/src/utils/store.ts
@@ -12,6 +12,10 @@ interface FacetInfo {
   label: string;
 }
 
+export interface ResultListInfo {
+  focusOnNextNewResult(): void;
+}
+
 interface FacetValueFormat<ValueType> {
   format(facetValue: ValueType): string;
   content?(facetValue: ValueType): VNode;
@@ -42,6 +46,8 @@ export type AtomicStore = {
    * This is a better indicator than only the "firstSearchExecuted" of the Headless search state.
    */
   loadingFlags: string[];
+  fieldsToInclude: string[];
+  resultList?: ResultListInfo;
 };
 
 export const initialStore: () => AtomicStore = () => ({
@@ -54,6 +60,7 @@ export const initialStore: () => AtomicStore = () => ({
   iconAssetsPath: '',
   mobileBreakpoint: DEFAULT_MOBILE_BREAKPOINT,
   loadingFlags: [],
+  fieldsToInclude: [],
 });
 
 export const registerFacetToStore = <T extends FacetType, U extends string>(
@@ -67,6 +74,13 @@ export const registerFacetToStore = <T extends FacetType, U extends string>(
 
   store.state[facetType][data.facetId] = data;
   store.state.facetElements.push(data.element);
+};
+
+export const registerResultListToStore = (
+  store: ObservableMap<AtomicStore>,
+  data: ResultListInfo
+) => {
+  store.set('resultList', data);
 };
 
 // https://terodox.tech/how-to-tell-if-an-element-is-in-the-dom-including-the-shadow-dom/