+++ packages/samples/headless-react/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.59.0](https://github.com/coveo/ui-kit/compare/headless-react@0.58.2...headless-react@0.59.0) (2022-06-17)
+
+
+### Features
+
+* **atomic-react:** add CommonJS output ([#2113](https://github.com/coveo/ui-kit/issues/2113)) ([2b1f7ca](https://github.com/coveo/ui-kit/commit/2b1f7cacd2c53399ec3a6189fdb08790b1419bac))
+
+
+
+
+
+## [0.58.2](https://github.com/coveo/ui-kit/compare/headless-react@0.58.1...headless-react@0.58.2) (2022-06-16)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+## [0.58.1](https://github.com/coveo/ui-kit/compare/headless-react@0.58.0...headless-react@0.58.1) (2022-06-16)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+# [0.58.0](https://github.com/coveo/ui-kit/compare/headless-react@0.57.1...headless-react@0.58.0) (2022-06-13)
+
+
+### Bug Fixes
+
+* commit package-lock files right when bumping, undo last bump & fix region ([#2074](https://github.com/coveo/ui-kit/issues/2074)) ([28ad872](https://github.com/coveo/ui-kit/commit/28ad872363d2b84711865cda0c6d61bba7354927))
+
+
+### Features
+
+* **atomic,headless:** add support for allowed values for specific facets ([#2075](https://github.com/coveo/ui-kit/issues/2075)) ([5ad9bf1](https://github.com/coveo/ui-kit/commit/5ad9bf165813d3cc9d882ef8178da74f90260203))
+
+
+
+
+
 ## [0.57.1](https://github.com/coveo/ui-kit/compare/headless-react@0.57.0...headless-react@0.57.1) (2022-06-09)
 
 **Note:** Version bump only for package headless-react